+++ src/app/util/project.js
@@ -4,10 +4,6 @@
 const isStartView = state =>
 	!state || state.view === 'start' || state.view == null;
 
-const isDemoProject = state =>
-	!!(state && state.project && state.project.id === 'demo');
-
 module.exports = {
-	isStartView,
-	isDemoProject
+	isStartView
 };